fix(types): Remove invalid context field from ValidationIssue
- ValidationIssue doesn't have a context field
- Remove the invalid keyword argument
- Keep all other required fields
- All pre-commit hooks now pass

+++ services/tracklist_service/src/services/cue_generation_service.py
@@ -209,7 +209,7 @@ class CueGenerationService:
             else:
                 failed_count += 1
 
-        processing_time = (datetime.now(timezone.utc) - start_time).total_seconds() * 1000
+        processing_time = int((datetime.now(timezone.utc) - start_time).total_seconds() * 1000)
 
         return BatchCueGenerationResponse(
             success=failed_count == 0,
